+++ autoload/hello.lua
@@ -1,0 +1,1 @@
+print('hello from autoload')
